requirements.txt
streamlit==1.30.0
PyPDF2==3.0.1
python-docx==1.1.0
langchain==0.1.16
langchain-groq==0.1.5
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.1
